Extract damage pipeline to rules/damage.py as pure functions (#595)
Move CombatEngine._apply_damage_modifiers and _apply_damage_adjustments
into dnd_engine/rules/damage.py as standalone pure functions, mirroring
the sibling AC-formula selector in rules/ac_formulas.py. The CombatEngine
methods become thin delegates, so existing combat behavior is unchanged.

This establishes the single canonical damage chokepoint that the
non-combat callers (condition-tick damage, auto-hit spells, thrown items)
can route through without depending on CombatEngine — the prerequisite
for closing the take_damage bypasses tracked in #595.

Behavior-preserving refactor: new unit suite for the rules function plus
the existing immunity/resistance/underwater suites all pass.
